Script for mapping STCN authors to Biografisch Portaal (tested, works)
Whenever a name-match is found, the script does an additional check for
at least 1 year match to ensure (at least minimally) that no matches
occur for different persons. This means that only authors in STCN that
have at least one year in their 'rdfs:label' will have a potential
match in the Bioportaal. We can do a manual scan, if need be, to create
matches between authors that do not have years associated with them.